move tests outside
