+++ __tests__/packageData.test.js
@@ -1,11 +1,11 @@
-const { updatePackageJSON } = require('../');
-const upgradeDeps = require('../upgradeDeps');
-const babelCoreFixture = require('../../fixtures/babel-core');
-const jestFixture = require('../../fixtures/jest');
-const depsFixture = require('../../fixtures/deps');
-const depsFixtureEarlierBeta = require('../../fixtures/deps-earlier-beta.json');
-const scriptsMochaFixture = require('../../fixtures/scripts-mocha');
-const scriptsBabelNodeFixture = require('../../fixtures/scripts-babel-node');
+const { updatePackageJSON } = require('../src/');
+const upgradeDeps = require('../src/upgradeDeps');
+const babelCoreFixture = require('../fixtures/babel-core');
+const jestFixture = require('../fixtures/jest');
+const depsFixture = require('../fixtures/deps');
+const depsFixtureEarlierBeta = require('../fixtures/deps-earlier-beta.json');
+const scriptsMochaFixture = require('../fixtures/scripts-mocha');
+const scriptsBabelNodeFixture = require('../fixtures/scripts-babel-node');
 
 test('packages', () => {
   expect(upgradeDeps(depsFixture, "7.0.0-beta.39")).toMatchSnapshot();
